+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.0 branch
 ### 2.0.4-RC4
 * spots are automatically played before video;
+* the duration of the video is saved in the database;
 * the Youtube API are used now.
 
 ### 2.0.3-RC3